update 2.0
I added a lot of stuff here. Spend like 2 hours just adding 4 items to each container. 3 containers. 1 wing 1 pizza and 1 drink. Added images and grid to each container. This will make it each for me to make everything responsive later. I added 3 buttons and added to cart button to the pizza container. I need to add the add to cart button to every item. Going to focus more on the javascript stuff now. I'll add the description of the item calorie and pricing later. Also, need to make a homepage showing off the "popular items" I want it to look clean and moving containers and animations since I never messed around with that.