Update 3.0
I added a lot of javascript on this commit. I just added some arrays with basic functions to do the math for the total order of the 4 pizzas you can select. I also worked out some of the functions for the tip that will take the total and let you pick a tip. I got that working in the console but not the HTML rn cause I want to add drinks and wings to the total options and then I'll add the tip you selected into the total. Same with The driver fee.