Update 6.0
Added a lot in the javascript dom. Spend 2 hours just going through what I want when you click order. I give the client 2 options to order on the card or with cash. The cash option displays some text and the credit card will display a lot of user inputs with options to type in them. I need to add a lot more to the credit card page by adding if-else statements to each input and also making the order button on that page work and will open another page saying the order has been completed. After that, I have some basic cool ideas I want to add. The main thing I will work on after all of that will be the homepage and responsive design. I have a lot of work for this week.